use int16_t for fk_tmev_t.expired

+++ fk_ev.h
@@ -39,7 +39,7 @@ typedef struct fk_tmev_s fk_tmev_t;
 struct fk_tmev_s {
     FK_HEAP_LEAF_HEADER;/* for heap */
 
-    int expired;/* whether in the expired list */
+    int16_t expired;/* whether in the expired list */
     uint8_t type;/* FK_TMEV_CYCLE|FK_TMEV_ONCE */
     uint32_t interval;/* milliseconds */
     struct timeval when;/* save the trigger point time: now + timeout */
